+++ internal/fleet/agentdownloadsource/acc_test.go
@@ -1,0 +1,127 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package agentdownloadsource_test
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/elastic/terraform-provider-elasticstack/internal/acctest"
+	"github.com/elastic/terraform-provider-elasticstack/internal/clients"
+	"github.com/elastic/terraform-provider-elasticstack/internal/clients/fleet"
+	"github.com/elastic/terraform-provider-elasticstack/internal/diagutil"
+	"github.com/elastic/terraform-provider-elasticstack/internal/versionutils"
+	"github.com/hashicorp/go-version"
+	sdkacctest "github.com/hashicorp/terraform-plugin-testing/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+var minVersionFleetAgentDownloadSource = version.Must(version.NewVersion("8.13.0"))
+
+func TestAccResourceFleetAgentDownloadSource(t *testing.T) {
+	random := sdkacctest.RandString(8)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		CheckDestroy:             checkResourceFleetAgentDownloadSourceDestroy,
+		ProtoV6ProviderFactories: acctest.Providers,
+		Steps: []resource.TestStep{
+			{
+				SkipFunc: versionutils.CheckIfVersionIsUnsupported(minVersionFleetAgentDownloadSource),
+				Config:   testAccResourceFleetAgentDownloadSourceCreate(random),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "name", fmt.Sprintf("Agent Download Source %s", random)),
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "default", "false"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "host", "https://artifacts.elastic.co/downloads/elastic-agent"),
+				),
+			},
+			{
+				SkipFunc: versionutils.CheckIfVersionIsUnsupported(minVersionFleetAgentDownloadSource),
+				Config:   testAccResourceFleetAgentDownloadSourceUpdate(random),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "name", fmt.Sprintf("Updated Agent Download Source %s", random)),
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "default", "false"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_agent_download_source.test", "host", "https://artifacts.elastic.co/downloads/elastic-agent"),
+				),
+			},
+			{
+				SkipFunc:          versionutils.CheckIfVersionIsUnsupported(minVersionFleetAgentDownloadSource),
+				Config:            testAccResourceFleetAgentDownloadSourceUpdate(random),
+				ResourceName:      "elasticstack_fleet_agent_download_source.test",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func testAccResourceFleetAgentDownloadSourceCreate(suffix string) string {
+	return fmt.Sprintf(`
+provider "elasticstack" {
+  kibana {}
+}
+
+resource "elasticstack_fleet_agent_download_source" "test" {
+  name      = "Agent Download Source %s"
+  default   = false
+  host      = "https://artifacts.elastic.co/downloads/elastic-agent"
+}
+`, suffix)
+}
+
+func testAccResourceFleetAgentDownloadSourceUpdate(suffix string) string {
+	return fmt.Sprintf(`
+provider "elasticstack" {
+  kibana {}
+}
+
+resource "elasticstack_fleet_agent_download_source" "test" {
+  name      = "Updated Agent Download Source %s"
+  default   = false
+  host      = "https://artifacts.elastic.co/downloads/elastic-agent"
+}
+`, suffix)
+}
+
+func checkResourceFleetAgentDownloadSourceDestroy(s *terraform.State) error {
+	client, err := clients.NewAcceptanceTestingClient()
+	if err != nil {
+		return err
+	}
+
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "elasticstack_fleet_agent_download_source" {
+			continue
+		}
+
+		fleetClient, err := client.GetFleetClient()
+		if err != nil {
+			return err
+		}
+		resp, diags := fleet.GetAgentDownloadSource(context.Background(), fleetClient, rs.Primary.ID, "")
+		if diags.HasError() {
+			return diagutil.FwDiagsAsError(diags)
+		}
+		if resp != nil && resp.JSON200 != nil {
+			return fmt.Errorf("fleet agent download source id=%v still exists, but it should have been removed", rs.Primary.ID)
+		}
+	}
+	return nil
+}